backend/requirements.txt
fastapi>=0.100.0
uvicorn
pydantic>=2.0
google-generativeai
